chore: regenerate unified extraction bindings

+++ packages/java/io/xberg/EmbeddingBackendBridge.java
@@ -228,7 +228,7 @@ public final class EmbeddingBackendBridge implements AutoCloseable {
             try (var nameArena = Arena.ofShared()) {
                 var nameCs = nameArena.allocateFrom(impl.name());
                 MemorySegment outErr = nameArena.allocate(ValueLayout.ADDRESS);
-                int rc = (int) NativeLib.XBERG_REGISTER_EMBEDDING_BACKEND.invoke(
+                int rc = (int) (long) NativeLib.XBERG_REGISTER_EMBEDDING_BACKEND.invoke(
                     nameCs,
                     bridge.vtableSegment(),
                     MemorySegment.NULL,
@@ -257,7 +257,7 @@ public final class EmbeddingBackendBridge implements AutoCloseable {
             try (var nameArena = Arena.ofShared()) {
                 var nameCs = nameArena.allocateFrom(name);
                 MemorySegment outErr = nameArena.allocate(ValueLayout.ADDRESS);
-                int rc = (int) NativeLib.XBERG_UNREGISTER_EMBEDDING_BACKEND.invoke(nameCs, outErr);
+                int rc = (int) (long) NativeLib.XBERG_UNREGISTER_EMBEDDING_BACKEND.invoke(nameCs, outErr);
                 if (rc != 0) {
                     MemorySegment errPtr = outErr.get(ValueLayout.ADDRESS, 0);
                     String msg = errPtr.equals(MemorySegment.NULL)
@@ -281,7 +281,7 @@ public final class EmbeddingBackendBridge implements AutoCloseable {
         try {
             try (var arena = Arena.ofShared()) {
                 MemorySegment outErr = arena.allocate(ValueLayout.ADDRESS);
-                int rc = (int) NativeLib.XBERG_CLEAR_EMBEDDING_BACKEND.invoke(outErr);
+                int rc = (int) (long) NativeLib.XBERG_CLEAR_EMBEDDING_BACKEND.invoke(outErr);
                 if (rc != 0) {
                     MemorySegment errPtr = outErr.get(ValueLayout.ADDRESS, 0);
                     String msg = errPtr.equals(MemorySegment.NULL)
